# Optimize Analyze Endpoint